requirements.txt
torch==2.8.0
torchvision==0.23.0
transformers==4.57.1
datasets==4.0.0
tokenizers==0.22.1
accelerate==1.2.0
deepspeed==0.16.7
decord==0.6.0
pillow==10.4.0
qwen-vl-utils==0.0.14
peft==0.14.0
bitsandbytes==0.45.0
flash-attn==2.8.3
verl==0.7.0.dev0
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.12.3
numpy==1.26.4
pandas==2.0.1
scipy==1.14.1
scikit-learn==1.6.0
evaluate==0.4.3
wandb==0.19.9
tensorboard==2.19.0
loguru==0.7.3
tqdm==4.67.1
pyyaml==6.0.2
requests==2.32.3
jsonlines==4.0.0
isort==5.13.2
langchain==0.3.10